+++ react/components/ShippingSimulator/components/ShippingTable.js
@@ -4,6 +4,7 @@ import { FormattedMessage } from 'react-intl'
 
 import ShippingTableRow from './ShippingTableRow'
 
+import styles from '../styles.css'
 export default class ShippingTable extends Component {
   static propTypes = {
     /** Shipping informations */
@@ -36,14 +37,15 @@ export default class ShippingTable extends Component {
       return (
         <FormattedMessage id="shipping.empty-sla">
           {text => (
-            <span className="vtex-shipping__no-shipping-message dib t-small mt4">{text}</span>
-          )}
-        </FormattedMessage>
+            <span className={`${styles.shippingNoMessage} dib t-small mt4`}>{text}</span>
+          )
+          }
+        </FormattedMessage >
       )
     }
 
     return (
-      <table className="vtex-shipping-table bt bb b--muted-4 c-muted-1 ph0 pv3 mt4 w-100">
+      <table className={`${styles.shippingTable} bt bb b--muted-4 c-muted-1 ph0 pv3 mt4 w-100`} >
         <tbody>
           {slaList.map(shipping => (
             <ShippingTableRow key={shipping.id} {...shipping} />