+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support to CSS modules in `CollectionBadges`.
 - Support to CSS modules in `DiscountBadge`.
 - Support to CSS modules in `AvailabilitySubscriber`.
+- Support to CSS modules in `Logo`.
+- Support to CSS modules in `ShippingSimulator`.
 
 ## [3.5.8] - 2019-01-15
 ### Fixed